06_Projects/cleanstat-infrastructure/requirements.txt
# ============================================================
# CLEANSTAT INFRASTRUCTURE - PYTHON DEPENDENCIES
# ============================================================

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Caching & Queue
redis==5.0.1
celery==5.3.6
kombu==5.3.5

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
eth-account==0.10.0
cryptography==41.0.7

# HTTP & API
httpx==0.26.0
requests==2.31.0

# File Processing
Pillow==10.2.0
python-magic==0.4.27

# IPFS
pinatapy==1.0.0
ipfshttpclient==0.8.0

# Blockchain
web3==6.11.3
eth-utils==2.1.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3

# Logging
structlog==24.1.0
sentry-sdk[fastapi]==1.40.0

# Monitoring
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.0.0

# Rate Limiting
slowapi==0.1.9

# Task Scheduling
apscheduler==3.10.4

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.26.0

# Code Quality
black==23.12.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# Frontend (Gradio)
gradio==4.0.0
